rag_based_youtube_chatbot/requirements.txt
# Core LangChain OSS
langchain==0.3.19
langchain-core==0.3.40
langchain-community==0.3.18
langchain-text-splitters==0.3.6

# Vector store
chromadb==0.6.2
langchain-chroma==0.2.2

# Hugging Face + transformers
langchain-huggingface==0.1.2
transformers==4.46.3
torch==2.4.1

# YouTube transcripts
youtube-transcript-api==1.2.3

# Env & web
python-dotenv==1.0.1
fastapi==0.115.5
uvicorn==0.32.0
